working version

+++ synchronize.c
@@ -2,19 +2,26 @@
 #include "commons.h"
 #include "utility.h"
 
+int isRecursive = 1;
 
 void deleteNotMatching(const char *srcPath, const char *destPath) {
     DIR *destDir = opendir(destPath);
     struct dirent *file;
 
     while ((file = readdir(destDir))) {
+    	if(strcmp(file->d_name, ".") == 0 || strcmp(file->d_name, "..") == 0)
+    		continue;
         const char *fileInSource = appendToPath(srcPath, file->d_name);
         const char *fileInDest = appendToPath(destPath, file->d_name);
         const int isDestFileDirectory = isDirectory(fileInDest);
 
-        if (!fileExists(fileInSource, isDestFileDirectory))
-            removeFile(fileInDest);
+        if (!fileExists(fileInSource, isDestFileDirectory)){
+        	syslog(LOG_INFO, "File removed %s", fileInDest);
+        	removeFile(fileInDest);
+        }
+            
     }
+    closedir(destDir);
 }
 
 void copyNotMatching(const char *srcPath, const char *destPath) {
@@ -22,6 +29,8 @@ void copyNotMatching(const char *srcPath, const char *destPath) {
     struct dirent *file;
 
     while ((file = readdir(srcDir))) {
+	    if(strcmp(file->d_name, ".") == 0 || strcmp(file->d_name, "..") == 0)
+    		continue;
         const char *fileInSource = appendToPath(srcPath, file->d_name);
         const char *fileInDest = appendToPath(destPath, file->d_name);
         const int isSourceFileDirectory = isDirectory(fileInSource);
@@ -30,15 +39,17 @@ void copyNotMatching(const char *srcPath, const char *destPath) {
             getDateOfModify(srcPath) < getDateOfModify(destPath)) {
 
             copyFile(fileInSource, fileInDest, isSourceFileDirectory);
+            syslog(LOG_INFO, "File created %s", fileInDest);
         }
 
-        if (isSourceFileDirectory && isRecursive)
-            sync(fileInSource, fileInDest);
+        if (isSourceFileDirectory && isDirectory(fileInDest) && isRecursive)
+            syncDirectories(fileInSource, fileInDest);
 
     }
+    closedir(srcDir);
 }
 
-void sync(const char *sourcePath, const char *destPath) {
+void syncDirectories(const char *sourcePath, const char *destPath) {
     //sleep for x seconds
 
     deleteNotMatching(sourcePath, destPath);
